+++ projects/igniteui-angular/src/lib/slider/thumb/thumb-slider.component.ts
@@ -1,5 +1,4 @@
 import {
-    NgModule,
     Component,
     Input,
     HostListener,
@@ -9,10 +8,11 @@ import {
     EventEmitter,
     OnInit,
     OnDestroy,
-    TemplateRef} from '@angular/core';
-import { Subject } from 'rxjs';
+    TemplateRef
+} from '@angular/core';
 import { takeUntil } from 'rxjs/operators';
 import { SliderHandle } from '../slider.common';
+import { Subject } from 'rxjs';
 
 /**
  * @hidden
@@ -23,9 +23,15 @@ import { SliderHandle } from '../slider.common';
 })
 export class IgxSliderThumbComponent implements OnInit, OnDestroy {
 
-    private _timer;
     private _isActiveLabel = false;
+    private _isPressed = false;
     private _destroy$ = new Subject<boolean>();
+
+    private get thumbPositionX() {
+        const thumbBounderies = this.nativeElement.getBoundingClientRect();
+        const thumbCenter = (thumbBounderies.right - thumbBounderies.left) / 2;
+        return thumbBounderies.left + thumbCenter;
+    }
 
     public isActive = false;
 
@@ -51,9 +57,6 @@ export class IgxSliderThumbComponent implements OnInit, OnDestroy {
     public step: number;
 
     @Input()
-    public fromHandler: boolean;
-
-    @Input()
     public templateRef: TemplateRef<any>;
 
     @Input()
@@ -68,44 +71,49 @@ export class IgxSliderThumbComponent implements OnInit, OnDestroy {
     @Output()
     public onChange = new EventEmitter<any>();
 
+    @Output()
+    public onHoverChange = new EventEmitter<boolean>();
+
     @HostBinding('attr.tabindex')
     public tabindex = 0;
 
+    @HostBinding('attr.z-index')
+    public zIndex = 0;
+
     @HostBinding('class.igx-slider__thumb-from')
     public get thumbFromClass() {
-        return this.fromHandler;
+        return this.type === SliderHandle.FROM;
     }
 
     @HostBinding('class.igx-slider__thumb-to')
     public get thumbToClass() {
-        return !this.fromHandler;
+        return this.type === SliderHandle.TO;
     }
 
     @HostBinding('class.igx-slider__thumb-from--active')
     public get thumbFromActiveClass() {
-        return this.fromHandler && this._isActiveLabel;
+        return this.type === SliderHandle.FROM && this._isActiveLabel;
     }
 
     @HostBinding('class.igx-slider__thumb-to--active')
     public get thumbToActiveClass() {
-        return !this.fromHandler && this._isActiveLabel;
+        return this.type === SliderHandle.TO && this._isActiveLabel;
+    }
+
+    @HostBinding('class.igx-slider__thumb--pressed')
+    public get thumbPressedClass() {
+        return this.isActive && this._isPressed;
     }
 
     public get nativeElement() {
         return this._elementRef.nativeElement;
     }
 
-    private get thumbPositionX() {
-        const thumbBounderies = this.nativeElement.getBoundingClientRect();
-        const thumbCenter = (thumbBounderies.right - thumbBounderies.left) / 2;
-        return thumbBounderies.left + thumbCenter;
-    }
-
     public get destroy(): Subject<boolean> {
         return this._destroy$;
     }
 
-    constructor (private _elementRef: ElementRef) { }
+    constructor(private _elementRef: ElementRef) { }
 
     /**
      * @hidden
@@ -124,6 +132,16 @@ export class IgxSliderThumbComponent implements OnInit, OnDestroy {
     public ngOnDestroy() {
         this._destroy$.next(true);
         this._destroy$.complete();
+    }
+
+    @HostListener('pointerenter')
+    public onPinterEnter() {
+        this.onHoverChange.emit(true);
+    }
+
+    @HostListener('pointerleave')
+    public onPointerLeave() {
+        this.onHoverChange.emit(false);
     }
 
     @HostListener('keydown', ['$event'])
@@ -148,43 +166,27 @@ export class IgxSliderThumbComponent implements OnInit, OnDestroy {
     @HostListener('blur')
     public onBlur() {
         this.isActive = false;
+        this.zIndex = 0;
     }
 
     @HostListener('focus')
     public onFocusListener() {
         this.isActive = true;
+        this.zIndex = 1;
     }
 
-    public showThumbLabel() {
-        if (this.disabled) {
-            return;
-        }
-
-        if (this.continuous) {
-            return;
-        }
-
-        if (this._timer !== null) {
-            clearTimeout(this._timer);
-        }
-
-        this._isActiveLabel = true;
+    /**
+     * Show thumb label and ripple.
+     */
+    public showThumbIndicators() {
+        this.toggleThumbIndicators(true);
     }
 
-
-    public hideThumbLabel() {
-        if (this.disabled) {
-            return;
-        }
-
-        if (this.continuous) {
-            return;
-        }
-
-        this._timer = setTimeout(
-            () => this._isActiveLabel = false,
-            this.thumbLabelVisibilityDuration
-        );
+    /**
+     * Hide thumb label and ripple.
+     */
+    public hideThumbIndicators() {
+        this.toggleThumbIndicators(false);
     }
 
     private updateThumbValue(mouseX: number) {
@@ -210,5 +212,13 @@ export class IgxSliderThumbComponent implements OnInit, OnDestroy {
 
     private stepToProceed(scaleX, stepDist) {
         return Math.round(scaleX / stepDist) * this.step;
+    }
+
+    private toggleThumbIndicators(visible: boolean) {
+        this._isPressed = visible;
+
+        if (!this.continuous) {
+            this._isActiveLabel = visible;
+        }
     }
 }